# eigen-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Eigen Backend Test",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
...
  "paths": {
    "/books": {
      "get": {
        "tags": ["Books"],
        "summary": "get all books in database",
        "operationId": "getAllBooks",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Book list fetched successfully"
                    },
                    "books": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string",
                            "example": "B001"
                          },
                          "title": {
                            "type": "string",
                            "example": "Pocahontas"
                          },
                          "author": {
                            "type": "string",
                            "example": "John Smith"
                          },
                          "stock": {
                            "type": "number",
                            "example": 5
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "post": {
        "tags": ["Books"],
        "summary": "saveBook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "code": "B009",
                  "title": "Testing Book",
                  "author": "Yusuf",
                  "stock": 10
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "Bad request response",
            "content": {
              "application/json": {}
            }
          },
          "409": {
            "description": "Conflict response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal server error response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "patch": {
        "tags": ["Books"],
        "summary": "updateBook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "code": "B009",
                  "title": "Testing Book",
                  "author": "Haha",
                  "stock": 10
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "Bad request response",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Error Not Found response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal server error response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  }
}